refactor: structure add: packages
